requirements.txt
fastapi==0.109.2
pydantic==1.10.13
pytest==8.0.0
Requests==2.31.0
uvicorn==0.27.1
pact-python==2.1.1